fix screensaver scale shoule be set float

+++ screensaver/main.cpp
@@ -74,9 +74,9 @@ int main(int argc, char *argv[])
         s.winId();
         s.windowHandle()->setParent(window);
         XGetWindowAttributes (QX11Info::display(), wid, &xwa);
-        int scale = 1;
+        float scale = 1;
         if(!qgetenv("QT_SCALE_FACTOR").isNull())
-            scale = qgetenv("QT_SCALE_FACTOR").toInt();
+            scale = qgetenv("QT_SCALE_FACTOR").toFloat();
 
         s.resize(xwa.width/scale,xwa.height/scale);
         s.move(0,0);
